Add SR_SUBSCR_OPER_ONCE option for oper subscr
This option causes a component subscribing to provide operational pull
data to be called no more than once per-subscription for a given get
on the operational datastore.

This can be used to optimise certain usecases. For example, if a
component is subscribed to provide pull state below a list, the default
behaviour causes a component to be called for each element in the list
(unless the request was for a specific element in the list).

As the number of elements in the parent list(s) grows it can take a
significant amount of time, proportional to the time required to actually
collect the data, to issue and handle a callback for each element when
data for all (or a large subset of) elements is requested.

This option therefore gives the capability to handle pull requests in
a "one shot" fashion where it is more performant to do so.