sysrepo: add SR_UNDERLYING_VALUE option
This option is provided for use with sr_get_items() to allow easy
retrieval of the underlying integer value of enums, rather than the
typical string representation.

An application could achieve this with a combination sr_get_data() and
then looking up the schema. However the intention of this option is to
provide a more convenient alternative, particularly in Rust applications
(where this would currently require unsafe code).