feature: out of band sysrepo inspection
- allow disabling connection liveliness check

If SYSREPO_DISABLE_CONN_LIVELINESS_CHECK is present in the environment,
assume that all sysrepo connections are alive without actually checking
them.

This allows collecting sysrepo SHM/state from one environment (eg. which
may be in a failed state), and inspecting it in a different environment.
Otherwise when we try to inspect the state in that other environment,
sysrepo will think connections are dead, and will clean up the state we
are trying to inspect.