Filter subscription notification at producer side
When a producer creates some changes for a certain module
sysrepo notifies all the subscriber threads, which then
filter based on the xpath, whether the change is relevant or not

By doing the filtering at the producer side, we can reduce the
overhead of notifying many threads and optimize resource use.

This should reduce the overhead of socket notification, locks
and improve commit times.

Add tests