cached module read locks info for srmon
we can cache the list of readers while we create the lyd_node for each
entry for reporting to sysrepo-monitoring.
This reduces the time for which locks are held.

Also, Ideally, `rwlock->writer` and `rwlock->upgr` should be read while
holding `rwlock->mutex`, so as to not get inconsistent values. Else, we may
have a scenario where a lock downgrades from writer to upgr, and we may read
both writer and upgr existing at the same time. But this should not
really cause any problems.

Closes: ENG-3840