include leaf-list value as XPath predicate
Revert upstream commit 97c6fabbd6c82c8643c0286e129635231472ee59 (and
make requisite updates to match intervening changes to the code)
which removed the predicate from leaf-list XPaths when generating
sr_val_t objects for return from sr_get_change_next(), purportedly
for compatibility with older versions of sysrepo.

We don't care about such compatibility and including the predicate
makes it easier to generically deal with changes in components where
we don't want to understand a specific schema.

For example, to delete a specific element from a leaf-list, the value
must be provided by predicate in the XPath passed to sr_delete_item(),
otherwise the entire list is deleted. This change allows the XPath
provided by a sr_val_t from sr_get_change_next() to be passed into
sr_delete_item() to remove that specific element, and without the
component needing to know that the XPath refers to a leaf-list (which
it cannot infer from other details in the sr_val_t, hence would
otherwise need to be aware of the schema).