implement sr_install_module_perm
The context is replaced everytime a new module is installed.

Which means that installing 100 modules will be extremely inefficient.
Providing a new API to install all modules in a directory with
permissions is a really fast way to install all the modules on bootup.

Ideally we should have the modules installed at image build time.