subscription memory check for validity
in libyang1, unless a reuse flag was passed, we always created a new
subscription. However in v2.x, the memory in the subscription context is
expected to be NULL first time.

If application passes an uninitialized pointer, and it is not zeroed,
then we can have undefined behavior.

To prevent this, a magic will be stored in the properly initialized
subscription context, and we will check for it when reusing an existing
context to add more subscriptions to it.