bugfix shm_sub event type finished instead of none
For change_sub events, listeners should use SR_SUB_EV_FINISHED instead
of SR_SUB_EV_NONE to signal that they are done. Only the originators of
the event should set it to SR_SUB_EV_NONE.

sr_shmsub_notify_new_wrlock waits for SR_SUB_EV_NONE, and it can assume
that the previous event is finished but it's originator may not have
seen it yet.

To prevent this, previously orig_cid was used, but this is not reliable.
If a timeout happens while the listeners have the read lock,
the originator cannot clear orig_cid, and the shm becomes unusable
after this.

Since event is an atomic type, we can always set it even though we don't
have a WRITE lock.

RPCs will continue to use SR_SUB_EV_NONE for SR_SUB_EV_NOTIF as the
originators don't always wait, and setting event to SR_SUB_EV_FINISHED
is not needed.

Closes: ENG-6332