commit-history-check to match upstream tag
It is important that the upstream tag exists in the current branch.
In case a rebase/squash/merge has changed the commit history,
we must detect it as early as possible.

Adding it to gbp.conf will detect in local builds.
graphnos-ci build have hooks disabled for good reason.

A new graphnos-ci job should prevent this from being merged.