add version check to debian/tests/install
It is good to add a version check to make sure that the debian/tests run
with the binaries built with this changeset rather than an older version
available in apt sources.

This will detect if debian/tests/control is not updated or improperly
updated after package name changes etc.