no discard oper push data on disconnect
During a system reboot or shutdown, it is not ideal to spend time
discarding operational data and informing subscribers of the deletes.

A supervisory process owning the reboot can signal this by calling the
API `sr_oper_data_no_discard_on_disconnect` to set a flag in the main shm.

After this, any subsequent processes disconnecting from sysrepo can do
so much faster.