sr_unsubscribe without change_sub WRITE lock
Support scrapping (safely) subscriptions in the ext SHM when WRITE
lock for change_sub is not available.

Background:
change_sub READ lock is held for the entire duration of a
`sr_changes_notify_store` operation.

Depending on the module, this can take a really long time, if a new
change is underway. A process that needs to stop (for process restart or
because the system needs to reboot) can suffer really long delays.

The situation is further worsened if notify is waiting for a process
that has recently crashed, or subscribers of multiple priorities exist.

Unsubscribing in a timely manner is also important so that the processes
can release their resources in destructors and not cause memory bugs.

Solution:
Although we cannot remove the subscription from ext SHM without a WRITE
change_sub lock, we can atomically mark the subscription as SCRAPPED.

Subsequent change_sub add to the same module and datastore will detect
this and clean up all `SCRAPPED` subs as it posseses WRITE lock.

From the perspective of publishers, a SCRAPPED subscription is the same
as a suspended subscription and are not notified of events.

Races:

1. If SCRAP happens while a publisher is waiting for it, the only way
to notify the publisher is by processing the event, as the publisher
does not track the xpath_ids it is waiting for and find scraps.

2. SCRAP after subscriber_count is determined by publisher,
but before event is notified.

If `sr_changes_notify_store` ongoing:

   a. before `subscriber_count` is determined => No action needed.

   b. before `evpipe` is notified => Correct subscriber_count to
   wait for, and skip waiting if there is no subscription notified.

   c. after `evpipe` is notified => Process the event.

Else => No action needed.

Since unsubscribe holds subscription context WRITE lock, event
processing cannot happen by the subscription handler thread during the
SCRAP. Once the call to SCRAP returns, we can be assured that the
subscription context can be freed up and the callback is never invoked
again.

Closes: ENG-6692