shutdown-mgr to disable oper discard on disconnect
optimize a system shutdown for sysrepo with a daemon to disable
oper discard on disconnect.

During a system reboot or shutdown, it is pointless to spend time
discarding operational data and informing subscribers of the deletes.

`sysrepo-shutdown-mgr` can be started after other daemons connect to
sysrepo and awaits SIGTERM and SIGINT signals.

When a system shutdown is initiated, this daemon could be stopped
before all other sysrepo applications.

When it is sent a SIGTERM or SIGINT, it sets a flag
in the main shm which does two things:
1. `sr_disconnect` skips `sr_discard_oper_changes`.
2. `sr_get_data` skips dead connection check.

So, any subsequent processes disconnecting from sysrepo will leave
their pushed operational data in the shared memory.
Any remaining readers will still see the old data, but this should not
cause any problems.

tests:
Added test for sysrepo-shutdown-mgr to oper import test script.

ENG-6495