update the SHM VERSION
When changes are made to the contents of shared memory, it is necessary
to increment the version, so older versions of the shm are not
mistakenly read by the newer library.

However using only the SR_SHM_VER can be confusing because upstream also
increments this.

Secondly, debian patch changes can be lost on intake of newer
code from upstream.

So, making SR_SHM_VER a combination of a SR_DEBIAN_SHM_VER and the
upstream SR_SHM_VER should ensure keeping the version correct.