add RefuseManual Start and Stop to shutdown-mgr
We do not want manual users to stop shutdown-mgr service inadvertently.

From the systemd unit manpage:
This is mostly a safety feature to ensure that the user does not
accidentally activate units that are not intended to be activated
explicitly, and not accidentally deactivate units that are not
intended to be deactivated.