shm_sub BUGFIX donot store error on lost_lock
When the originator cannot obtain the WRITE lock back for an event,
because another process (a subscriber) has a READ lock etc. lost_lock is
set. After this, no abort event is sent by the originator.

In this case, we must not set the event to error as orig_cid is already
cleared and can make the shm unusable as the error will never be
cleared.

Closes: ENG-6746