shm change_sub has_sub must account for dead CIDs
When a subscriber is dead and discovered, we update mod->xpath_id_count
among other fields.

Usually, before calling sr_shmsub_change_notify_has_subscription we
check for mod->xpath_id_count != 0
But in sr_shmsub_change_notify_change, has_sub is called twice in
succession without a check.

To prevent this, we should check for this in the function.

Also simplify update xpath_ids logic for deleting dead CID.

- tests
It is difficult to trigger the dead CID discovery without injecting some
sleep in the source code. So this area will remain a gap for now.

Closes: ENG-7009