always attempt recovery on delayed rwlock acquire
With every retry attempt of acquiring a rwlock, we can also check the
aliveness of the CID. A dead CID has the ability to cause system-wide
delays until the lock is recovered.

This mainly entails doing a CID liveness check on all lock holders.
The current mechanism for liveness check involves checking if a file is
locked by the owning process. A faster and better mechanism can be
implemented but is not trivial and from profiling for the performance
impact, the current mechanism only incurs a max penalty of < 10 us each
iteration.

Closes: ENG-6984