notify subscribers to discard-items changes
Merge the "discard-items" nodes with data in the operational datastore.
This will allow for subscribers to be notified when the data owner
discards some data.
However, this now has the side-effect of losing the pushed data. That is
simply replacing the "discard-items" node in a future edit_batch will
not restore the already discareded data.

Do not persist sysrepo discard-items nodes into the SHM in all cases.
When there is a single session pushing data into a module, there is no
need to store all discard-items data into the SHM.

For certain types of data, such as a list with timestamp as the key, the
application may keep pushing newer data, and discarding older data
periodically.

In this case, the operational data SHM will grow unbounded, causing
memory and CPU usage to increase over time.

While it is possible to manually edit the pushed operational changes
with libyang API, it is not convenient, and makes migrating to sysrepo
version 3.x painful.

A new API is defined sr_session_discard_items(), which operates just
like sr_discard_items but does not store the discard-items nodes in the
SHM, and has 2 limitations
1. This only affects pushed changes of this session and no other
   sessions.
2. This cannot be used to affect enabled data (running datastore data
   copied due to an active subscription)