ds-plugins - cache operational data when possible
push operational data needs to be loaded by the same connection that
pushed it originally. This is usually for updating the stored edit.

This happens by calling the load_cb() datastore plugin, which needs to
load the data from storage and constructed into a libyang tree.

This can be optimized by caching the last store operation, and reusing
the same data.

This cache becomes invalid when the libyang context changes, so it must
be discarded when yang modules are changed in any way.