oper ds ordering of session push data
 ### operational changes order
For each module, sessions can set the order explicitly by calling
`sr_set_oper_changes_order()`, or the next higher order is assigned to
it when it first pushes data to that module.

If a lower order process restarts, it will push data with
a higher order, unless it explicitly sets it's order.

 ### discard-items
Sessions can also push opaque "discard-items" nodes in addition to the
regular data nodes for the module.

These "discard-items" nodes can be used to specify XPaths that will not
be visible in the final observable "view" of the operational datastore.
Discard is *only* performed on *lower order* operational data.
They do not affect push data of sessions of higher order.

 ### `sr_set_oper_changes_order()`
1. Should be called by all push oper sessions of the module, if any
   sessions use "discard-items" matching that module.
2. As this function does not notify changes to subscribers, Order cannot
   be changed if the session has push data currently.
3. A session cannot reliably set order for all modules. This function
   must be called explicitly for each module before pushing operataionl
   data.

 ### Conclusion
For "discard-items" to work correctly on push operational data, all
the sessions should explicitly call `sr_set_oper_changes_order()` before
pushing any operational data for that module.