modinfo OPTIMIZE create ds_diff when needed
A ds_diff is not used by the json datastore plugin. It only uses the
module data to write out the entire contents each time.

A ds diff is not required by any datastore plugin if all data is being
deleted.

So, in both these cases, skip ds_diff creation, which can be very
expensive.